JETTY-713 Expose additional AbstractConnector methods via MBean

+++ jetty/branches/jetty-6.1/modules/jetty/src/main/java/org/mortbay/jetty/AbstractConnector.java
@@ -31,7 +31,7 @@ import org.mortbay.util.ajax.WaitingContinuation;
 
 
 /** Abstract Connector implementation.
- * This abstract implemenation of the Connector interface provides:<ul>
+ * This abstract implementation of the Connector interface provides:<ul>
  * <li>AbstractLifeCycle implementation</li>
  * <li>Implementations for connector getters and setters</li>
  * <li>Buffer management</li>
@@ -223,7 +223,7 @@ public abstract class AbstractConnector extends AbstractBuffers implements Conne
     /**
      * @return Returns the soLingerTime.
      */
-    public long getSoLingerTime()
+    public int getSoLingerTime()
     {
         return _soLingerTime;
     }
